#2887 Parametrized pool Medium/High Priority: - created class: CreateObjectUtils; - corrected CreateObjectUtils.parseObjects for empty strings;
